Clarify when single space-separated line is accepted
If it not accepted: Remove the mention of single line in the description
of the input, leaving only "multi-line". And always parse as if it is a
multi-line input.
If it is accepted: Update the parsing to check for newline characters
before actually parsing the input. Also add "(space-separated)" to
describe a multi-valued single-line input.