Update to ruff (instead of pylint (and isort))
Remove pylint as a dependency.
Remove all "# pylint: disable=..." statements.
Remove pylint and isort as pre-commit hooks.
Add ruff as a pre-commit hook.
Avoid running pylint in the CI.
Add a 'ruff' CI job.

Add (back) disable E501 statements.
E501: Line too long.
Add it for either lines with string-formatted type (which has to be on
one line) or an f-string with a single variable (i.e., a single-line
f-string part with an implemented logic for a variable that results in
the line being too long).