Change publish_on_pypi default
Add new upload_distribution input to match a new step in the publish
distribution job, where the distribution folder is uploaded as an
artifact.
Add new build_dir input to specify the build target folder for the
distribution.

Make publish_on_pypi a required input, in the hope that this will catch
users that do not set the input and rely on its default, which has now
changed and would result in successful workflows that would however not
upload anything to PyPI as expected.
Remove the requirement of publish_on_pypi in the next next version.

Update the documentation, adding warnings about the default value change
to both the top of the specific workflow page, but also to the landing
page.
Add an example on how to still use the reusable/callable workflow and
also the new PyPI Trusted Publisher feature.